requirements.txt
blis==0.7.11
catalogue==2.0.10
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
confection==0.1.5
cymem==2.0.11
en_ner_bc5cdr_md @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.1/en_ner_bc5cdr_md-0.5.1.tar.gz#sha256=7852184affa649bcc1417d11fb1075b99f483e1f8329d91ca296de8020a46a31
idna==3.10
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
MarkupSafe==3.0.2
murmurhash==1.0.12
numpy==1.23.5
packaging==25.0
pathlib_abc==0.1.1
pathy==0.11.0
preshed==3.0.9
pydantic==1.10.22
python-dotenv==1.1.0
requests==2.32.3
smart-open==6.4.0
spacy==3.4.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
thinc==8.1.10
tqdm==4.67.1
typer==0.7.0
typing_extensions==4.13.2
urllib3==2.4.0
wasabi==0.10.1
